fix(proxy) use the global proxy options
these options must be global since the https auth is used when
connecting, and the http one only when requesting.
So these settings span the process of connecting and requesting,
and hence we should not pass them into 'connect', because we would
have to ensure that the user would pass the same ones to
'request'.