fix(proxy) set proxy-auth when making request
the connect method cannot deal with the auth header for a http request,
only for https. Hence we store the value and inject is when making the
request. We need it on connect, to construct the proper pool-name.

We honor the existing request header if it is there, but that is a
bad idea, since the poolname is based on the one provided to
connect. So users should set the proxy options, and make requests
without that header to be safe.