Documentation is clearer about using same connection for multiple requests via 'request' function
If a previous HTTP response had a response body, and the body was never
read, then a subsequent call to 'request' using the same connection will
fail. This is because the unread response body will still be sitting in
the TCP socket buffers. As a result, when the subsequent call to
'request' tries to read the first line of its response to parse out the
HTTP version and status code, it will get the first line of the unread
response body instead, which will almost certainly not be a valid
HTTP status code line.

The documentation was not at all clear about this; make it explicitly
state what the user needs to do.